+++ shared/src/Datadog.Trace.ClrProfiler.Native/crashhandler.h
@@ -1,0 +1,24 @@
+#pragma once
+
+#include "util.h"
+#include "werapi.h"
+
+namespace datadog::shared::nativeloader
+{
+    struct WerContext
+    {
+        WCHAR* Environ;
+        int32_t EnvironLength;
+    };
+
+    class CrashHandler
+    {
+    public:
+        bool Register();
+        bool Unregister();
+
+    private:
+        std::wstring _crashHandler;
+        WerContext _context;
+    };
+}
